backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
requests==2.32.3
pypdf==4.2.0
python-docx==1.1.2
